+++ examples/mo_gaal_example.py
@@ -47,8 +47,14 @@ if __name__ == "__main__":
     y_test_pred = clf.predict(X_test)  # outlier labels (0 or 1)
     y_test_scores = clf.decision_function(X_test)  # outlier scores
 
+    # Assuming clf is an instance of your model
+    probabilities, confidence = clf.predict_proba(X_test, return_confidence=True)
+    # print("Probabilities shape:", probabilities.shape)
+    # print("Confidence shape:", confidence.shape)
+
     # evaluate and print the results
     print("\nOn Training Data:")
     evaluate_print(clf_name, y_train, y_train_scores)
     print("\nOn Test Data:")
     evaluate_print(clf_name, y_test, y_test_scores)
+
